TalentScan/requirements.txt
Flask
pdfminer.six
docx2txt